src/third_party/blink/web_tests/external/wpt/css/cssom-view/MediaQueryList-extends-EventTarget-interop-expected.txt

This is a testharness.js-based test.
PASS dispatchEvent triggers listener added with addListener
PASS listener added with addListener and addEventListener is called once
PASS listener added with addListener and addEventListener (capture) is called twice
PASS removeEventListener removes listener added with addListener
PASS removeEventListener (capture) doesn't remove listener added with addListener
PASS removeListener removes listener added with addEventListener
PASS removeListener doesn't remove listener added with addEventListener (capture)
FAIL listeners are called in order they were added, ignoring capture parameter assert_array_equals: triggerMQLEvent property 0, expected "addEventListener" but got "addListener"
Harness: the test ran to completion.
